Fix intersection subtyping; fix garbage `≤refinement≥` and intersection bases on Scala 2; fix union type constructor; fix summoning for higher-kinded constructors on Dotty; add additional tests for intersections, annotations and refinements; move all tests from Scala 2 LightTypeTag into SharedLightTypeTagTest
