+++ izumi-reflect/izumi-reflect/src/main/scala/izumi/reflect/DebugProperties.scala
@@ -28,7 +28,7 @@ object DebugProperties {
     * Set system property `-Dizumi.reflect.rtti.optimized.equals=false` to disable optimized `equals` comparison for
     * monomorphic [[izumi.reflect.macrortti.LightTypeTag]]s (instances of [[izumi.reflect.macrortti.LightTypeTag.ParsedLightTypeTag110]]).
     * Try this if you're experiencing "impossible" false negatives. While we believe that such false negatives are impossible,
-    * even between tags generated by
+    * even between tags generated by different versions of `izumi-reflect`, they could still be possible in actuality.
     *
     * {{{
     *   sbt -Dizumi.reflect.rtti.optimized.equals=false
