+++ izumi-reflect/izumi-reflect/src/main/scala_3/izumi/reflect/TagMacro.scala
@@ -1,71 +1,68 @@
 package izumi.reflect
 
-import scala.quoted.*
+import scala.quoted.{Expr, Quotes, Type}
 
 import izumi.reflect.macrortti.LightTypeTag
-import izumi.reflect.dottyreflection.Inspect
+import izumi.reflect.dottyreflection.{Inspect, InspectorBase}
 
 object TagMacro {
-
-  implicit inline def tag[A <: AnyKind]: Tag[A] = ${ createTagExpr[A] }
-
   def createTagExpr[A <: AnyKind: Type](using Quotes): Expr[Tag[A]] =
     new TagMacro().createTagExpr[A]
 }
 
-class TagMacro(using val ctx: Quotes) {
-  import ctx.reflect.*
+final class TagMacro(using override val qctx: Quotes) extends InspectorBase {
+  import qctx.reflect._
+
+  override def shift: Int = 0
 
   def createTagExpr[A <: AnyKind: Type]: Expr[Tag[A]] = {
-    val typeRepr = TypeRepr.of[A].dealias
-    if (allPartsStrong(typeRepr)) '{ Tag(classOf[Any], Inspect.inspect[A]) }
-    else {
-      val result = mkTag(typeRepr)
-      result.asInstanceOf[Expr[Tag[A]]]
+    val typeRepr = TypeRepr.of[A]
+    if (allPartsStrong(typeRepr)) {
+      // NB: this should always work, but currently causes `TestModel$::ApplePaymentProvider is not a type lambda, it cannot be parameterized` test failure
+      // There's probably a missing case where a higher-kinded type is not a type lambda, with splicing inbetween causing fixup by the compiler
+      //   val ltt = Inspect.inspectAny[A]
+      val ltt = '{ Inspect.inspect[A] }
+      '{ Tag[A](classOf[Any], $ltt) }
+    } else {
+      summonCombinedTag[A](typeRepr)
     }
   }
 
   private def mkLtt(typeRepr: TypeRepr): Expr[LightTypeTag] =
     if (allPartsStrong(typeRepr))
       typeRepr.asType match {
-        case '[a] => '{ Inspect.inspect[a] }
+//        case given Type[a] => Inspect.inspectAny[a]
+        case given Type[a] => '{ Inspect.inspect[a] }
       }
     else {
-      val result = mkTag(typeRepr)
+      val result = summonCombinedTag(typeRepr)
       '{ $result.tag }
     }
 
-  private def mkTag(typeRepr: TypeRepr): Expr[Tag[?]] =
+  private def summonCombinedTag[T <: AnyKind: Type](typeRepr: TypeRepr): Expr[Tag[T]] =
     typeRepr.dealias match {
 
       case x if x.typeSymbol.isTypeParam => summonTag(x)
 
-      case AppliedType(tpe, args) =>
-        val argsTags = Expr.ofList[LightTypeTag](args.map(arg => mkLtt(arg)))
-
-        tpe.asType match {
-          case '[a] =>
-            val tag = '{ Tag(classOf[Any], Inspect.inspect[a]) }
-            '{ Tag.appliedTag($tag, $argsTags) }
-        }
+      case AppliedType(ctor, args) =>
+        val ctorTag = createTagExpr(ctor.asType)
+        val argsTags = Expr.ofList(args.map(mkLtt))
+        '{ Tag.appliedTag[T]($ctorTag, $argsTags) }
 
       case andType: AndType =>
-        val ltts0: List[Expr[LightTypeTag]] = flattenAnd(andType).map(tt => mkLtt(tt))
+        val ltts0: List[Expr[LightTypeTag]] = flattenAnd(andType).map(mkLtt)
         val ltts: Expr[List[LightTypeTag]] = Expr.ofList(ltts0)
         andType.asType match {
           case '[a] =>
-            val struct = '{ Inspect.inspect[a] }
-            '{ Tag.refinedTag(classOf[Any], $ltts, $struct, Map.empty) }
+//            val structLtt = Inspect.inspectAny[a]
+            val structLtt = '{ Inspect.inspect[a] }
+            '{ Tag.refinedTag[T](classOf[Any], $ltts, $structLtt, Map.empty) }
         }
 
       case orType: OrType =>
-        val ltts0: List[Expr[LightTypeTag]] = flattenOr(orType).map(tt => mkLtt(tt))
+        val ltts0: List[Expr[LightTypeTag]] = flattenOr(orType).map(mkLtt)
         val ltts: Expr[List[LightTypeTag]] = Expr.ofList(ltts0)
-        orType.asType match {
-          case '[a] =>
-            val tag = '{ Tag(classOf[Any], Inspect.inspect[a]) }
-            '{ Tag.unionTag($tag, $ltts) }
-        }
+        '{ Tag.unionTag[T](classOf[Any], $ltts) }
 
       case _ =>
         throw new Exception(s"Unsupported type: $typeRepr")
@@ -83,11 +80,11 @@ class TagMacro(using val ctx: Quotes) {
       case _ => List(tpe)
     }
 
-  private def summonTag(typeRepr: TypeRepr): Expr[Tag[?]] =
+  private def summonTag[any <: AnyKind](typeRepr: TypeRepr): Expr[Tag[any]] =
     typeRepr.asType match {
-      case '[a] =>
+      case given Type[a] =>
         val message = s"Cannot find implicit Tag[${Type.show[a]}]"
-        Expr.summon[Tag[a]].getOrElse(report.errorAndAbort(message))
+        Expr.summon[Tag[a]].getOrElse(report.errorAndAbort(message)).asInstanceOf[Expr[Tag[any]]]
     }
 
   /**
@@ -106,7 +103,9 @@ class TagMacro(using val ctx: Quotes) {
       case NoPrefix() => true
       case TypeBounds(lo, hi) => allPartsStrong(lo) && allPartsStrong(hi)
       case TypeLambda(_, _, body) => allPartsStrong(body)
-      case _: TypeParamClause => true
+      case strange =>
+        log(s"Got unknown type component when checking strength: $strange")
+        true
     }
 
 }
