+++ izumi-reflect/izumi-reflect/src/main/scala_3/izumi/reflect/dottyreflection/InspectorBase.scala
@@ -9,6 +9,7 @@ trait InspectorBase {
 
   protected def shift: Int
 
+  // FIXME reimplement TrivialMacroLogger on Scala 3
   inline val debug = false
   type debug = debug.type
