+++ izumi-reflect/izumi-reflect/src/test/scala_2/izumi/reflect/test/IsDotty.scala
@@ -1,0 +1,9 @@
+package izumi.reflect.test
+
+import izumi.reflect.internal.fundamentals.platform.language.unused
+
+import scala.language.implicitConversions
+
+object IsDotty {
+  implicit def isDotty(@unused t: this.type): Boolean = false
+}
