+++ izumi-reflect/izumi-reflect/src/test/scala_3/izumi/reflect/test/IsDotty.scala
@@ -1,0 +1,3 @@
+package izumi.reflect.test
+
+inline val IsDotty = true
